refactor: 사이드바, 헤더에서 사용한은 icon을 public 폴더로 분리
