+++ public/icon/MainLogo.tsx
